remove the redundant format
