+++ src/fn/index.ts
@@ -11,7 +11,6 @@ export { qfp } from "./qfp"
 export { mlp } from "./mlp"
 export { ssop } from "./ssop"
 export { tssop } from "./tssop"
-export { sot236 } from "./sot236"
 export { sot363 } from "./sot363"
 export { sot23 } from "./sot23"
 export { dfn } from "./dfn"
@@ -26,7 +25,6 @@ export { pushbutton } from "./pushbutton"
 export { stampboard } from "./stampboard"
 export { stampreceiver } from "./stampreceiver"
 export { lqfp } from "./lqfp"
-export { sot235 } from "./sot235"
 export { breakoutheaders } from "./breakoutheaders"
 export { hc49 } from "./hc49"
 export { pad } from "./pad"
