+++ src/footprinter.ts
@@ -58,12 +58,10 @@ export type Footprinter = {
   axial: () => FootprinterParamsBuilder<"p" | "id" | "od">
   hc49: () => FootprinterParamsBuilder<"p" | "id" | "od" | "w" | "h">
   to220: () => FootprinterParamsBuilder<"w" | "h" | "p" | "id" | "od">
-  sot235: () => FootprinterParamsBuilder<"h" | "p" | "pl" | "pw">
   sot363: () => FootprinterParamsBuilder<"w" | "p" | "pl" | "pw">
   sot563: () => FootprinterParamsBuilder<"w" | "p" | "pl" | "pw">
   sot723: () => FootprinterParamsBuilder<"w" | "h" | "pl" | "pw">
   sot23: () => FootprinterParamsBuilder<"w" | "h" | "pl" | "pw">
-  sot236: () => FootprinterParamsBuilder<"w" | "p">
   sot89: () => FootprinterParamsBuilder<"w" | "p" | "pl">
   smc: () => FootprinterParamsBuilder<"w" | "h" | "p" | "pw" | "pl">
   minimelf: () => FootprinterParamsBuilder<"w" | "h" | "p" | "pw" | "pl">
