+++ tests/sot23.test.ts
@@ -15,7 +15,22 @@ test("sot23_w3_h1.5_p0.95mm", () => {
     "sot23_w3_h1.5_p0.95mm",
   )
 })
-test("sot23_num_pins5", () => {
+test("sot23_3", () => {
+  const circuitJson = fp.string("sot23_3").circuitJson()
+
+  const smtpad = circuitJson.filter((e) => e.type === "pcb_smtpad")
+  expect(smtpad).toBeDefined()
+  expect(smtpad.length).toBe(3)
+
+  const svgContent = convertCircuitJsonToPcbSvg(circuitJson)
+  expect(svgContent).toMatchSvgSnapshot(
+    import.meta.path,
+    "sot23_3",
+  )
+})
+
+
+test("sot23_5", () => {
   const circuitJson = fp.string("sot23_5").circuitJson()
 
   const smtpad = circuitJson.filter((e) => e.type === "pcb_smtpad")
@@ -24,4 +39,18 @@ test("sot23_num_pins5", () => {
 
   const svgContent = convertCircuitJsonToPcbSvg(circuitJson)
   expect(svgContent).toMatchSvgSnapshot(import.meta.path, "sot23_5")
+})
+
+test("sot23_6", () => {
+  const circuitJson = fp.string("sot23_6").circuitJson()
+
+  const smtpad = circuitJson.filter((e) => e.type === "pcb_smtpad")
+  expect(smtpad).toBeDefined()
+  expect(smtpad.length).toBe(6)
+
+  const svgContent = convertCircuitJsonToPcbSvg(circuitJson)
+  expect(svgContent).toMatchSvgSnapshot(
+    import.meta.path,
+    "sot23_6",
+  )
 })
